requirements.txt
flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
sqlalchemy==2.0.23
pydantic==2.5.3
gunicorn==21.2.0
flask-talisman==1.1.0
flask-limiter==3.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0